+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - added option to toggle extension UI components in readonly server mode
 - added reloadTasks function to Extension API for reloading tasks from disk
 - added support for saving edited prompts
+- added onTaskDeleted extension hook
 
 ## [0.76.0]